backend/requirements.txt
# ── Core ──
pydantic>=2.5.0,<2.11
pyyaml>=6.0
rank-bm25>=0.2.2

# ── Embeddings / Rerank ──
sentence-transformers>=3.0.0,<4.0
FlagEmbedding>=1.2.10
# huggingface-hub 版本由 transformers / sentence-transformers 自动约束，无需显式指定
tokenizers>=0.20.0

# ── LLM ──
# transformers >=4.48 强制要求 torch>=2.6（CVE-2025-32434）
# 若无法装 torch>=2.6，降级到 transformers==4.47.1 可绕过
transformers>=4.48.0,<5.0
openai>=1.30.0

# ── Vector DB ──
pymilvus>=2.4.6,<3.0

# ── Persistence ──
sqlalchemy>=2.0.0

# ── Parsing（多模态文档解析）──
pymupdf>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0,<13.0

# ── OCR（可选，PDF/图片解析用）──
pytesseract>=0.3.10
# paddleocr>=2.7.0            # 备选 OCR 引擎

# ── ASR 语音识别（视频处理需要）──
faster-whisper>=1.0.0

# ── API 服务 ──
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.7       # FastAPI 文件上传 (UploadFile/Form) 隐式依赖

# ── 工具库 ──
numpy>=1.24.0,<3.0
pandas>=2.0.0,<3.0
tqdm>=4.60.0
httpx>=0.27.0
orjson>=3.9.0
regex>=2023.0
jieba>=0.42.1
